sentinel/requirements.txt
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0
cdk-nag==2.28.0
dacite==1.8.1
packaging==23.2